logout処理を切り出し、とりあえず Terminal logout
